[CLEANUP] Traits for creation and modification dates
Also make the affected properties in the Administrator model
protected instead of private so that the static code analysis
will not mark them as unused.
